feat: Moved hc_report_builder and hc_session (with hc_spdx) into hc_analysis
