+++ libs/hc_analysis/Cargo.toml
@@ -7,16 +7,24 @@ publish = false
 
 [dependencies]
 content_inspector = "0.2.4"
+dirs = "5.0.1"
+dotenv = "0.15.0"
+curl = "0.4.38"
+glob = "0.3.0"
 hc_common = { path = "../hc_common" }
-hc_data = { path = "../hc_data" }
 hc_config = { path = "../hc_config" }
+hc_data = { path = "../hc_data" }
 hc_report = { path = "../hc_report" }
+hc_serde = { path = "../hc_serde" }
 hc_shell = { path = "../hc_shell" }
+hc_version = { path = "../hc_version" }
 maplit = "1.0.2"
 petgraph = { version = "0.6.0" }
+serde_derive = "1.0.137"
+spdx-rs = "0.5.0"
+toml = "0.8.12"
 unicode-normalization = "0.1.19"
 unicode-segmentation = "1.9.0"
+ureq = { version = "2.9.7", features = ["json"] }
 walkdir = "2"
-
-[dev-dependencies]
-hc_common = { path = "../hc_common" }
+xml-rs = "0.8"
