+++ libs/hc_analysis/src/lib.rs
@@ -2,7 +2,9 @@
 
 pub mod analysis;
 pub mod metric;
+pub mod report_builder;
 pub mod score;
+pub mod session;
 
 pub use analysis::{AnalysisProvider, AnalysisProviderStorage};
 pub use metric::{MetricProvider, MetricProviderStorage};
