+++ libs/hc_analysis/src/report_builder.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use hc_analysis::{analysis::AnalysisReport, score::ScoringResults};
+use crate::{analysis::AnalysisReport, score::ScoringResults, session::Session};
 use hc_common::{
 	error::{Error, Result},
 	hc_error, log,
@@ -8,7 +8,6 @@ use hc_common::{
 use hc_config::RiskConfigQuery;
 use hc_data::source::SourceQuery;
 pub use hc_report::*;
-use hc_session::Session;
 use std::default::Default;
 use std::result::Result as StdResult;
 
@@ -41,7 +40,7 @@ pub fn build_report(session: &Session, scoring: &ScoringResults) -> Result<Repor
 				_ => {
 					return Err(hc_error!(
 						"phase name does not match {} analysis",
-						hc_analysis::score::ACTIVITY_PHASE
+						crate::score::ACTIVITY_PHASE
 					))
 				}
 			}
@@ -73,7 +72,7 @@ pub fn build_report(session: &Session, scoring: &ScoringResults) -> Result<Repor
 				_ => {
 					return Err(hc_error!(
 						"phase name does not match {} analysis",
-						hc_analysis::score::AFFILIATION_PHASE
+						crate::score::AFFILIATION_PHASE
 					))
 				}
 			}
@@ -105,7 +104,7 @@ pub fn build_report(session: &Session, scoring: &ScoringResults) -> Result<Repor
 				_ => {
 					return Err(hc_error!(
 						"phase name does not match {} analysis",
-						hc_analysis::score::BINARY_PHASE
+						crate::score::BINARY_PHASE
 					))
 				}
 			}
@@ -136,7 +135,7 @@ pub fn build_report(session: &Session, scoring: &ScoringResults) -> Result<Repor
 				_ => {
 					return Err(hc_error!(
 						"phase name does not match {} analysis",
-						hc_analysis::score::CHURN_PHASE
+						crate::score::CHURN_PHASE
 					))
 				}
 			}
@@ -167,7 +166,7 @@ pub fn build_report(session: &Session, scoring: &ScoringResults) -> Result<Repor
 				_ => {
 					return Err(hc_error!(
 						"phase name does not match {} analysis",
-						hc_analysis::score::ENTROPY_PHASE
+						crate::score::ENTROPY_PHASE
 					))
 				}
 			}
@@ -198,7 +197,7 @@ pub fn build_report(session: &Session, scoring: &ScoringResults) -> Result<Repor
 				_ => {
 					return Err(hc_error!(
 						"phase name does not match {} analysis",
-						hc_analysis::score::IDENTITY_PHASE
+						crate::score::IDENTITY_PHASE
 					))
 				}
 			}
@@ -226,7 +225,7 @@ pub fn build_report(session: &Session, scoring: &ScoringResults) -> Result<Repor
 				_ => {
 					return Err(hc_error!(
 						"phase name does not match {} analysis",
-						hc_analysis::score::FUZZ_PHASE
+						crate::score::FUZZ_PHASE
 					))
 				}
 			}
@@ -257,7 +256,7 @@ pub fn build_report(session: &Session, scoring: &ScoringResults) -> Result<Repor
 				_ => {
 					return Err(hc_error!(
 						"phase name does not match {} analysis",
-						hc_analysis::score::REVIEW_PHASE
+						crate::score::REVIEW_PHASE
 					))
 				}
 			}
@@ -288,7 +287,7 @@ pub fn build_report(session: &Session, scoring: &ScoringResults) -> Result<Repor
 				_ => {
 					return Err(hc_error!(
 						"phase name does not match {} analysis",
-						hc_analysis::score::TYPO_PHASE
+						crate::score::TYPO_PHASE
 					))
 				}
 			}
@@ -483,7 +482,7 @@ pub fn build_pr_report(session: &Session, scoring: &ScoringResults) -> Result<Pr
 				_ => {
 					return Err(hc_error!(
 						"phase name does not match {} analysis",
-						hc_analysis::score::PR_AFFILIATION_PHASE
+						crate::score::PR_AFFILIATION_PHASE
 					))
 				}
 			}
@@ -517,7 +516,7 @@ pub fn build_pr_report(session: &Session, scoring: &ScoringResults) -> Result<Pr
 				_ => {
 					return Err(hc_error!(
 						"phase name does not match {} analysis",
-						hc_analysis::score::PR_CONTRIBUTOR_TRUST_PHASE
+						crate::score::PR_CONTRIBUTOR_TRUST_PHASE
 					))
 				}
 			}
@@ -551,7 +550,7 @@ pub fn build_pr_report(session: &Session, scoring: &ScoringResults) -> Result<Pr
 				_ => {
 					return Err(hc_error!(
 						"phase name does not match {} analysis",
-						hc_analysis::score::PR_MODULE_CONTRIBUTORS_PHASE
+						crate::score::PR_MODULE_CONTRIBUTORS_PHASE
 					))
 				}
 			}
