+++ libs/hc_analysis/src/session/spdx.rs
