+++ libs/hc_core/Cargo.toml
@@ -8,8 +8,6 @@ publish = false
 [dependencies]
 hc_analysis = { path = "../hc_analysis" }
 hc_common = { path = "../hc_common" }
-hc_report_builder = { path = "../hc_report_builder" }
-hc_session = { path = "../hc_session" }
 hc_shell = { path = "../hc_shell" }
 hc_version = { path = "../hc_version" }
