--- libs/hc_report_builder/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "hc_report_builder"
-version = "0.1.0"
-edition = "2021"
-license = "Apache-2.0"
-publish = false
-
-[dependencies]
-hc_analysis = { path = "../hc_analysis" }
-hc_common = { path = "../hc_common" }
-hc_config = { path = "../hc_config" }
-hc_data = { path = "../hc_data" }
-hc_report = { path = "../hc_report" }
-hc_session = { path = "../hc_session" }
