--- libs/hc_session/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "hc_session"
-version = "0.1.0"
-edition = "2021"
-license = "Apache-2.0"
-publish = false
-
-[dependencies]
-dirs = "5.0.1"
-dotenv = "0.15.0"
-curl = "0.4.38"
-glob = "0.3.0"
-hc_analysis = { path = "../hc_analysis" }
-hc_common = { path = "../hc_common" }
-hc_config = { path = "../hc_config" }
-hc_data = { path = "../hc_data" }
-hc_report = { path = "../hc_report" }
-hc_serde = { path = "../hc_serde" }
-hc_shell = { path = "../hc_shell" }
-hc_spdx = { path = "../hc_spdx" }
-hc_version = { path = "../hc_version" }
-serde_derive = "1.0.137"
-toml = "0.8.12"
-ureq = { version = "2.9.7", features = ["json"] }
-xml-rs = "0.8"
-
-[dev_dependencies]
-tempfile = "3.10.1"
